Renamed files to xe_*
* Removed l2 workspace alignment
