+++ include/cutlass/gemm/kernel/gemm_universal.hpp
@@ -65,6 +65,6 @@ struct IsCutlass3ArrayKernel<ProblemShape, cute::void_t<typename ProblemShape::U
 
 #if defined(SYCL_INTEL_TARGET)
 #include "cutlass/gemm/kernel/intel_pvc_gemm.hpp"
-#include "cutlass/gemm/kernel/intel_pvc_gemm_cooperative.hpp"
+#include "cutlass/gemm/kernel/xe_gemm_cooperative.hpp"
 #endif
 ////////////////////////////////////////////////////////////////////////////////
